# openApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BEJS Chapter 5 challenge",
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://localhost:2001/api/v1/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/transactions": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Post transaction",
        "description": "Post transaction with source account, destination account, and amount are required",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "source_account_id": {
                    "type": "integer"
                  },
                  "destination_account_id": {
                    "type": "integer"
                  },
                  "amount": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Transaction successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "source_account_id": {
                          "type": "integer"
                        },
                        "destination_account_id": {
                          "type": "integer"
                        },
                        "amount": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "enum": [
                        "Transaction successful"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad syntax"
          },
          "404": {
            "description": "Destination account/Source account not found"
          }
        }
      },
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Find all transactions",
        "description": "Returns all transactions",
        "responses": {
          "200": {
            "description": "get all transactions successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "wrapped": true
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "source_account_id": {
                            "type": "integer"
                          },
                          "destination_account_id": {
                            "type": "integer"
                          },
                          "amount": {
                            "type": "integer"
                          }
                        }
                      },
                      "message": {
                        "enum": [
                          "fetch transactions data success"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad syntax"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Find transaction by an id",
        "description": "Returns an object contains transaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of transaction to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get transaction success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "source_account_id": {
                          "type": "integer"
                        },
                        "destination_account_id": {
                          "type": "integer"
                        },
                        "amount": {
                          "type": "integer"
                        },
                        "source_account_info": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        },
                        "destination_account_info": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "message": {
                      "enum": [
                        "fetch transaction data success"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID must be a number (Bad Syntax Error)"
          },
          "404": {
            "description": "Transaction does not exist"
          }
        }
      },
      "delete": {
        "tags": [
          "transactions"
        ],
        "summary": "Delete a transaction by an id",
        "description": "Delete an object contains transaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of transaction to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete transaction success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "enum": [
                        "Transaction deleted"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID must be a number / Bad Syntax Error"
          },
          "404": {
            "description": "Transaction does not exist"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/accounts": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Post accounts",
        "description": "Post accounts with bank name, bank account number, and balance are required",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bank_name": {
                    "type": "string"
                  },
                  "bank_account_number": {
                    "type": "string"
                  },
                  "balance": {
                    "type": "string"
                  },
                  "user_id": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Create account successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "bank_name": {
                          "type": "string"
                        },
                        "bank_account_number": {
                          "type": "string"
                        },
                        "balance": {
                          "type": "string"
                        },
                        "user_id": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "enum": [
                        "Account created successfully"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Syntax Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "enum": [
                        "User_id must be a number"
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "enum": [
                        "Account number should be unique"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "enum": [
                        "User not exist"
                      ]
                    },
                    "message2": {
                      "enum": [
                        "Account not found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Find all accounts",
        "description": "Get all accounts object",
        "responses": {
          "200": {
            "description": "get all transactions successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "wrapped": true
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "bank_name": {
                            "type": "string"
                          },
                          "bank_account_number": {
                            "type": "string"
                          },
                          "balance": {
                            "type": "string"
                          },
                          "user_id": {
                            "type": "integer"
                          }
                        }
                      },
                      "message": {
                        "enum": [
                          "Get all accounts success"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Find an account details by id",
        "description": "Returns an account object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of transaction to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get transaction by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "bank_name": {
                          "type": "string"
                        },
                        "bank_account_number": {
                          "type": "string"
                        },
                        "balance": {
                          "type": "string"
                        },
                        "user_id": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "enum": [
                        "Fetch account data success"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Syntax Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "enum": [
                        "Account ID must be a number"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "enum": [
                        "Account not found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Delete an account",
        "description": "Delete an account object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of transaction to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get transaction by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "enum": [
                        "Account deleted"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Syntax Error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}